Refactor testlib
