+++ sdks/python/tests/testlib/__init__.py
@@ -1,10 +1,14 @@
-from .fake_message_processor import FakeMessageProcessor
-from .testlib_dsl import SpanModel, TraceModel, ANY_BUT_NONE, assert_traces_match
+from .backend_emulator_message_processor import BackendEmulatorMessageProcessor
+from .models import SpanModel, TraceModel
+from .assert_helpers import assert_dicts_equal, prepare_difference_report, assert_equal
+from .any_but_none import ANY_BUT_NONE
 
 __all__ = [
     "SpanModel",
     "TraceModel",
     "ANY_BUT_NONE",
-    "assert_traces_match",
-    "FakeMessageProcessor",
+    "assert_equal",
+    "assert_dicts_equal",
+    "prepare_difference_report",
+    "BackendEmulatorMessageProcessor",
 ]
